PER-702: Pan-Left/-Right Icon-State should not be enabled when it should be disabled anymore. Was cause by premature rounding-down/casting a double to an int.

+++ invesdwin-context-client-parent/invesdwin-context-client-swing-jfreechart/src/main/java/de/invesdwin/context/client/swing/jfreechart/panel/helper/PlotZoomHelper.java
@@ -415,7 +415,7 @@ public class PlotZoomHelper {
                 range = array[i].beforeLimitRange(range, rangeChanged);
             }
         }
-        final int length = (int) range.getLength();
+        final double length = range.getLength();
         final int gap = chartPanel.getAllowedMaximumRangeGap(length);
         final List<? extends TimeRangedOHLCDataItem> data = chartPanel.getMasterDataset().getData();
         final double minLowerBound = getMinLowerBoundWithGap(data, gap);
